ヘルプ画面の音声ライブラリの利用規約セクションのリデザイン (#1831)
* 音声ライブラリの利用規約セクションを新しいデザインへ変更

* Update src/components/template/HelpLibraryPolicySection.vue

* Update src/components/template/HelpLibraryPolicySection.vue

---------
